refactor(docs): consolidate logic
